requirements.txt
django >=4.0.0,<4.1.1
djangorestframework
pyyaml
requests
django-cors-headers